% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{nobs.pathtree}
\alias{nobs.pathtree}
\title{Number of Observations Used to Fit a Pathtree}
\usage{
\method{nobs}{pathtree}(object, ...)
}
\arguments{
\item{object}{A \code{pathtree}.}

\item{...}{Ignored.}
}
\value{
Integer. Number of state observations used to fit the tree.
}
\description{
Number of Observations Used to Fit a Pathtree
}
