% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{pathway_exists}
\alias{pathway_exists}
\title{Test Whether a Pathway Exists in the Tree}
\usage{
pathway_exists(tree, pathway)
}
\arguments{
\item{tree}{A \code{pathtree}.}

\item{pathway}{Character. Pathway as arrow-notation string or
character vector.}
}
\value{
Logical scalar.
}
\description{
Test Whether a Pathway Exists in the Tree
}
\examples{
\donttest{
tr <- context_tree(matrix(sample(c("A","B"), 50, TRUE), 5),
                   max_depth = 2L, nmin = 1L)
pathway_exists(tr, "A")
}
}
