% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_tree.R
\name{print.transitiontrees_group}
\alias{print.transitiontrees_group}
\title{Print a Group of Context Trees}
\usage{
\method{print}{transitiontrees_group}(x, ...)
}
\arguments{
\item{x}{A \code{transitiontrees_group} (named list of \code{transitiontrees}s).}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a Group of Context Trees
}
