% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{print.transitiontrees_tune}
\alias{print.transitiontrees_tune}
\title{Print a Pathtree Tuning Grid}
\usage{
\method{print}{transitiontrees_tune}(x, n = 10L, ...)
}
\arguments{
\item{x}{A \code{transitiontrees_tune} object.}

\item{n}{Integer. Number of configurations to print. Default 10.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a Pathtree Tuning Grid
}
