% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.transitrees_bootstrap}
\alias{print.transitrees_bootstrap}
\title{Print a Pathtree Bootstrap}
\usage{
\method{print}{transitrees_bootstrap}(x, n = 10L, digits = 3L, ...)
}
\arguments{
\item{x}{A \code{transitrees_bootstrap} object.}

\item{n}{Integer. Number of top pathways to print. Default 10.}

\item{digits}{Integer. Numeric digits for the printed table.
Default 3.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a Pathtree Bootstrap
}
