% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_tree.R
\name{print.transitrees_group}
\alias{print.transitrees_group}
\title{Print a Group of Context Trees}
\usage{
\method{print}{transitrees_group}(x, ...)
}
\arguments{
\item{x}{A \code{transitrees_group} (named list of \code{transitrees}s).}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a Group of Context Trees
}
