% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitiontrees-package.R
\docType{package}
\name{transitiontrees-package}
\alias{transitiontrees}
\alias{transitiontrees-package}
\title{transitiontrees: Context Trees and Variable-Order Markov Models}
\description{
Fits prediction suffix trees (Ron, Singer & Tishby 1996) and
variable-order Markov models from categorical sequence data.
Provides multiple pruning criteria, per-context Kullback-Leibler
diagnostics, smoothing, prediction, and tree visualisation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mohsaqr/transitiontrees}
  \item Report bugs at \url{https://github.com/mohsaqr/transitiontrees/issues}
}

}
\author{
\strong{Maintainer}: Mohammed Saqr \email{saqr@saqr.me}

}
\keyword{internal}
