% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitrees-package.R
\docType{package}
\name{transitrees-package}
\alias{transitrees}
\alias{transitrees-package}
\title{transitrees: Context Trees and Variable-Order Markov Models}
\description{
Fits prediction suffix trees (Ron, Singer & Tishby 1996) and
variable-order Markov models from categorical sequence data.
Provides multiple pruning criteria, per-context Kullback-Leibler
diagnostics, smoothing, prediction, and tree visualisation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mohsaqr/pathtree}
  \item Report bugs at \url{https://github.com/mohsaqr/pathtree/issues}
}

}
\author{
\strong{Maintainer}: Mohammed Saqr \email{saqr@saqr.me}

}
\keyword{internal}
